move auth inteceptor to stanalone class
